# src/__generated__/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Record_string.string-or-string-Array_": {
				"properties": {},
				"additionalProperties": {
					"anyOf": [
						{
							"type": "string"
						},
						{
							"items": {
								"type": "string"
							},
							"type": "array"
						}
					]
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
...
			"BaseResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"errors": {
						"$ref": "#/components/schemas/Record_string.string-or-string-Array_"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"OrderValidatorCode": {
				"description": "Error errors returned by the order validator contract",
				"enum": [
					0,
					101,
					111,
					112,
					113,
					201,
					211,
					212,
					213,
					301,
					311,
					312,
					321,
					322,
					401,
					402,
					411,
					412,
					413,
					414,
					415,
					421,
					422,
					501,
					502,
					503,
					601,
					611,
					612,
					621,
					622,
					623,
					631,
					632,
					633,
					634,
					701,
					702,
					801,
					802,
					901,
					902
				],
				"type": "number"
			},
			"Pick_Order.Exclude_keyofOrder.createdAt-or-id-or-invalidated-or-validator_codes__": {
				"properties": {},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_Order.createdAt-or-id-or-invalidated-or-validator_codes_": {
				"$ref": "#/components/schemas/Pick_Order.Exclude_keyofOrder.createdAt-or-id-or-invalidated-or-validator_codes__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"CreateOrderRequest": {
				"properties": {
					"signature": {
						"type": "string"
					},
					"chainId": {
						"type": "number",
						"format": "double"
					},
					"quoteType": {
						"type": "number",
						"format": "double"
					},
					"globalNonce": {
						"type": "string"
					},
					"subsetNonce": {
						"type": "number",
						"format": "double"
					},
					"orderNonce": {
						"type": "string"
					},
					"strategyId": {
						"type": "number",
						"format": "double"
					},
					"collectionType": {
						"type": "number",
						"format": "double"
					},
					"collection": {
						"type": "string"
					},
					"currency": {
						"type": "string"
					},
					"signer": {
						"type": "string"
					},
					"startTime": {
						"type": "number",
						"format": "double"
					},
					"endTime": {
						"type": "number",
						"format": "double"
					},
					"price": {
						"type": "string"
					},
					"itemIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"amounts": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"additionalParameters": {
						"type": "string"
					}
				},
				"required": [
					"signature",
					"chainId",
					"quoteType",
					"globalNonce",
					"subsetNonce",
					"orderNonce",
					"strategyId",
					"collectionType",
					"collection",
					"currency",
					"signer",
					"startTime",
					"endTime",
					"price",
					"itemIds",
					"amounts",
					"additionalParameters"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateOrderNonceRequest": {
				"properties": {
					"address": {
						"type": "string"
					},
					"chainId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"address",
					"chainId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidateOrderRequest": {
				"properties": {
					"tokenIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"chainId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"tokenIds",
					"chainId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "Hypercerts API",
		"version": "1.0.0",
		"description": "API for uploading allow lists and hypercert metadata. Visit /graphql for the GraphQL API.",
		"contact": {
			"url": "https://hypercerts.org"
		}
	},
	"paths": {
...
		"/v1/marketplace/orders": {
			"post": {
				"operationId": "StoreOrder",
				"responses": {
					"201": {
						"description": "Order created successfully",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean"
												}
											},
											"required": [
												"error",
												"data",
												"message",
												"success"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {},
												"data": {
													"properties": {
														"order": {
															"$ref": "#/components/schemas/Omit_Order.createdAt-or-id-or-invalidated-or-validator_codes_"
														},
														"validatorCodes": {
															"items": {
																"$ref": "#/components/schemas/OrderValidatorCode"
															},
															"type": "array"
														},
														"valid": {
															"type": "boolean"
														},
														"id": {
															"type": "string"
														}
													},
													"required": [
														"order",
														"validatorCodes",
														"valid",
														"id"
													],
													"type": "object"
												},
												"success": {
													"type": "boolean"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"success",
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {},
												"data": {
													"properties": {
														"validator_codes": {
															"items": {
																"type": "number",
																"format": "double"
															},
															"type": "array"
														},
														"invalidated": {
															"type": "boolean"
														},
														"id": {
															"type": "string"
														},
														"hypercert_id": {
															"type": "string"
														},
														"createdAt": {
															"type": "string"
														},
														"additionalParameters": {
															"type": "string"
														},
														"price": {
															"type": "string"
														},
														"endTime": {
															"type": "number",
															"format": "double"
														},
														"startTime": {
															"type": "number",
															"format": "double"
														},
														"signer": {
															"type": "string"
														},
														"currency": {
															"type": "string"
														},
														"collection": {
															"type": "string"
														},
														"collectionType": {
															"type": "number",
															"format": "double"
														},
														"strategyId": {
															"type": "number",
															"format": "double"
														},
														"orderNonce": {
															"type": "string"
														},
														"subsetNonce": {
															"type": "number",
															"format": "double"
														},
														"globalNonce": {
															"type": "string"
														},
														"quoteType": {
															"type": "number",
															"format": "double"
														},
														"chainId": {
															"type": "number",
															"format": "double"
														},
														"signature": {
															"type": "string"
														},
														"hash": {
															"type": "string"
														},
														"status": {
															"type": "string"
														},
														"amounts": {
															"items": {
																"type": "number",
																"format": "double"
															},
															"type": "array"
														},
														"itemIds": {
															"items": {
																"type": "string"
															},
															"type": "array"
														}
													},
													"required": [
														"validator_codes",
														"invalidated",
														"id",
														"hypercert_id",
														"createdAt",
														"additionalParameters",
														"price",
														"endTime",
														"startTime",
														"signer",
														"currency",
														"collection",
														"collectionType",
														"strategyId",
														"orderNonce",
														"subsetNonce",
														"globalNonce",
														"quoteType",
														"chainId",
														"signature",
														"hash",
														"status",
														"amounts",
														"itemIds"
													],
													"type": "object"
												},
												"success": {
													"type": "boolean"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"success",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"422": {
						"description": "Unprocessable content",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"message": "Order could not be created"
										}
									}
								}
							}
						}
					}
				},
				"description": "Submits a new order for validation and storage on the database.",
				"tags": [
					"Marketplace"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateOrderRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteOrder",
				"responses": {
					"200": {
						"description": "Order deleted successfully",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"422": {
						"description": "Unprocessable content",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"message": "Order could not be deleted"
										}
									}
								}
							}
						}
					}
				},
				"description": "Delete order from database",
				"tags": [
					"Marketplace"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"signature": {
										"type": "string"
									},
									"orderId": {
										"type": "string"
									}
								},
								"required": [
									"signature",
									"orderId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/v1/marketplace/order-nonce": {
			"post": {
				"operationId": "UpdateOrderNonce",
				"responses": {
					"200": {
						"description": "Order nonce updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"properties": {
												"nonce_counter": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"nonce_counter"
											],
											"type": "object"
										},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"422": {
						"description": "Unprocessable content",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"message": "Order nonce could not be updated"
										}
									}
								}
							}
						}
					}
				},
				"description": "Updates and returns the order nonce for a user on a specific chain.",
				"tags": [
					"Marketplace"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateOrderNonceRequest"
							}
						}
					}
				}
			}
		},
		"/v1/marketplace/orders/validate": {
			"post": {
				"operationId": "ValidateOrder",
				"responses": {
					"200": {
						"description": "Order validated successfully",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"items": {
												"properties": {
													"validator_codes": {
														"items": {
															"type": "number",
															"format": "double"
														},
														"type": "array"
													},
													"invalidated": {
														"type": "boolean"
													},
													"id": {
														"type": "string"
													},
													"hypercert_id": {
														"type": "string"
													},
													"createdAt": {
														"type": "string"
													},
													"additionalParameters": {
														"type": "string"
													},
													"amounts": {
														"items": {
															"type": "number",
															"format": "double"
														},
														"type": "array"
													},
													"itemIds": {
														"items": {
															"type": "string"
														},
														"type": "array"
													},
													"price": {
														"type": "string"
													},
													"endTime": {
														"type": "number",
														"format": "double"
													},
													"startTime": {
														"type": "number",
														"format": "double"
													},
													"signer": {
														"type": "string"
													},
													"currency": {
														"type": "string"
													},
													"collection": {
														"type": "string"
													},
													"collectionType": {
														"type": "number",
														"format": "double"
													},
													"strategyId": {
														"type": "number",
														"format": "double"
													},
													"orderNonce": {
														"type": "string"
													},
													"subsetNonce": {
														"type": "number",
														"format": "double"
													},
													"globalNonce": {
														"type": "string"
													},
													"quoteType": {
														"type": "number",
														"format": "double"
													},
													"chainId": {
														"type": "number",
														"format": "double"
													},
													"signature": {
														"type": "string"
													}
												},
												"type": "object"
											},
											"type": "array"
										},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"422": {
						"description": "Unprocessable content",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"message": "Order could not be validated"
										}
									}
								}
							}
						}
					}
				},
				"description": "Validates an order and marks it as invalid if validation fails.",
				"tags": [
					"Marketplace"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ValidateOrderRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}